refactor(vault): Rename ManagedClient to CachedClient and consolidate cache logic
- Renamed ManagedClient to CachedClient throughout the codebase for clarity
- Renamed createGroup singleflight to authGroup to better reflect its purpose
- Consolidated cache_key.go into client_pool_cache.go to reduce file sprawl
- Removed standalone cache_key files after merging logic
- All tests continue to pass without modification

This refactoring improves code organization while maintaining full backward
compatibility and test coverage.
